mcp/grafana/dashboardview/requirements.txt
pydantic
mcp
python-dotenv
fastapi
uvicorn
fastmcp
requests